refactor: move fetch_lemma_by_user_query to search.py
