+++ CreeDictionary/tests/API_tests/model_test.py
@@ -5,7 +5,8 @@ from typing import List
 import pytest
 from hypothesis import assume, given
 
-from API.models import Wordform, fetch_lemma_by_user_query
+from API.models import Wordform
+from API.search import fetch_lemma_by_user_query
 from CreeDictionary import settings
 from paradigm import EmptyRowType, InflectionCell, Layout, TitleRow
 from tests.conftest import lemmas
